# specs/001-starter-app-template/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Starter App Template API
  description: API specification for the Next.js starter app template with Supabase, Stripe, and RAG integrations
  version: 1.0.0
  contact:
    name: Starter Template Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-app.vercel.app/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT token
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        avatar_url:
          type: string
          format: uri
        subscription_status:
          type: string
          enum: [free, active, past_due, canceled, trialing]
        theme_preference:
          type: string
          enum: [light, dark, system]
        onboarding_completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
...
security:
  - BearerAuth: []
paths:
...
  /user/profile:
    get:
      summary: Get user profile
      description: Retrieve the authenticated user's profile information
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      summary: Update user profile
      description: Update user profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  maxLength: 50
                last_name:
                  type: string
                  maxLength: 50
                theme_preference:
                  type: string
                  enum: [light, dark, system]
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /stripe/webhook:
    post:
      summary: Stripe webhook endpoint
      description: Handle Stripe webhook events
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Stripe webhook event
      responses:
        '200':
          description: Webhook processed successfully
        '400':
          description: Invalid webhook signature or payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...